Updated several examples.
Updated several examples, added CMakeLists.txt files for cmake builds, and fixed some warning with convert_to from Numbers.h. Modernized getting DPI in some instances.

+++ examples/COLORS2/COLORS2.cpp
@@ -6,7 +6,7 @@ HRESULT Colors2::Initialize()
 	//get app title from resource file.
 	std::tstring apptitle;
     apptitle.resize(256);
-    LoadString(HINST_THISCOMPONENT,IDS_APP_TITLE,apptitle.data(),apptitle.length());
+    LoadString(HINST_THISCOMPONENT,IDS_APP_TITLE,apptitle.data(),static_cast<int>(apptitle.length()));
     apptitle.shrink_to_fit();
 
 	hr = ::IsWindow(Create(nullptr, &Window::rcDefault, apptitle.c_str(), IDC_COLORS2)) ? S_OK : E_FAIL;
@@ -27,7 +27,6 @@ void Colors2::OnExit(UINT uNotifyCode, int nID, Window wndCtl)
 
 void Colors2::OnAbout(UINT uNotifyCode, int nID, Window wndCtl)
 {
-
 	if(!about)
 	{
 		if(about.DoModal() == IDOK)
@@ -35,7 +34,6 @@ void Colors2::OnAbout(UINT uNotifyCode, int nID, Window wndCtl)
 			MessageBox::Show(TEXT("OK pressed!"), TEXT("WinTest"),MessageBoxButtons::Ok, MessageBoxIcon::Information);
 		}
 	}
-
 }
 
 void Colors2::OnClose()
@@ -45,9 +43,7 @@ void Colors2::OnClose()
 
 int Colors2::OnCreate(LPCREATESTRUCT lpCreateStruct)
 {
-	//additional window initalization would go here.
-
-	
+	//additional window initalization would go here.	
 	if(scrolldlg.Create(*this))
 		scrolldlg.ShowWindowAsync(ShowWindowType::Normal);
     return 0;
